refactor: TmdbApi.ts 파일 경로 변경
